Feature - Error Handler 404

+++ src/controllers/customer-address.controller.ts
@@ -8,7 +8,7 @@ export default class CustomerAddressController{
 
     get = async(req: any, res: any, next: any) => {
         try {
-            var data = await this.repository.getAll()
+            var data = await this.repository.getAll();
             res.status(200).send(data)
           } catch (erro) {
             res.status(500).send({
@@ -20,7 +20,13 @@ export default class CustomerAddressController{
     getById = async(req: any, res: any, next: any) => {
         let id = req.params.id
         try {
-            var data = await this.repository.getById(id)
+            var data = await this.repository.getById(id);
+            if(data === null){
+              res.status(404).send({
+                 error_code: 'CUSTOMER ADDRESS NOT FOUND',
+                 error_description: 'Customer Address not found'
+              })
+            }
             res.status(200).send(data)
           } catch (erro) {
             res.status(500).send({
@@ -44,7 +50,14 @@ export default class CustomerAddressController{
     put = async(req: any, res: any, next: any) => {
         let id = req.params.id
         try {
-            var data = await this.repository.update(id, req.body)
+            var data = await this.repository.update(id, req.body);
+            if(data[0] === 0){
+              res.status(404).send({
+                 error_code: 'CUSTOMER ADDRESS NOT FOUND',
+                 error_description: 'Customer Address not found'
+              })
+              return
+            }
             res.status(200).send(data)
           } catch (erro) {
             res.status(500).send({
@@ -56,9 +69,17 @@ export default class CustomerAddressController{
     delete = async(req: any, res: any, next: any) => {
         let id = req.params.id
         try {
+            const data = await this.repository.getById(id);
+            if(data == null){
+              res.status(404).send({
+                error_code: 'CUSTOMER ADDRESS NOT FOUND',
+                error_description: 'Customer Address not found'
+             })
+             return
+            }
             await this.repository.deleteById(id)
             res.status(200).send({
-               message: 'Deleted'
+               message: 'Deleted'+data
             })
           } catch (erro) {
             res.status(500).send({
